Füge JUnit und erste Tests hinzu

+++ src/main/java/Calculator.java
@@ -11,10 +11,10 @@ public class Calculator {
         return a * b;
     }
 
-    public double divide(int a, int b) {
+    public int divide(int a, int b) {
         if (b == 0) {
             throw new IllegalArgumentException("Cannot divide by zero.");
         }
-        return (double) a / b;
+        return  a / b;
     }
 }
